refactor: remove logger di modules
