+++ features/resume/internal-api/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/resume/internalapi/ProveYourIdentityEntryPoints.kt
@@ -4,8 +4,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.Immutable
 import androidx.compose.ui.Modifier
 import com.squareup.anvil.annotations.ContributesTo
-import uk.gov.onelogin.criorchestrator.libraries.di.scopes.ActivityScope
-import uk.gov.onelogin.criorchestrator.libraries.di.scopes.CriOrchestratorScope
+import uk.gov.onelogin.criorchestrator.libraries.di.ActivityScope
+import uk.gov.onelogin.criorchestrator.libraries.di.CriOrchestratorScope
 
 @Immutable
 fun interface ProveYourIdentityEntryPoints {
