+++ features/resume/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/resume/internal/ProveYourIdentityEntryPointsImpl.kt
@@ -9,8 +9,8 @@ import com.squareup.anvil.annotations.ContributesBinding
 import uk.gov.onelogin.criorchestrator.features.resume.internal.root.ProveYourIdentityRoot
 import uk.gov.onelogin.criorchestrator.features.resume.internal.root.ProveYourIdentityViewModelModule
 import uk.gov.onelogin.criorchestrator.features.resume.internalapi.ProveYourIdentityEntryPoints
-import uk.gov.onelogin.criorchestrator.libraries.di.scopes.ActivityScope
-import uk.gov.onelogin.criorchestrator.libraries.di.scopes.CriOrchestratorScope
+import uk.gov.onelogin.criorchestrator.libraries.di.ActivityScope
+import uk.gov.onelogin.criorchestrator.libraries.di.CriOrchestratorScope
 import javax.inject.Inject
 import javax.inject.Named
