+++ libraries/di/src/main/kotlin/uk/gov/onelogin/criorchestrator/libraries/di/CriOrchestratorScope.kt
@@ -1,4 +1,4 @@
-package uk.gov.onelogin.criorchestrator.libraries.di.scopes
+package uk.gov.onelogin.criorchestrator.libraries.di
 
 /**
  * Anvil scope used to help Anvil make a connection between the Dagger component and which Dagger
