+++ sdk/shared-api/src/main/kotlin/uk/gov/onelogin/criorchestrator/sdk/sharedapi/CriOrchestratorComponent.kt
@@ -1,7 +1,7 @@
 package uk.gov.onelogin.criorchestrator.sdk.sharedapi
 
 import com.squareup.anvil.annotations.ContributesTo
-import uk.gov.onelogin.criorchestrator.libraries.di.scopes.CriOrchestratorScope
+import uk.gov.onelogin.criorchestrator.libraries.di.CriOrchestratorScope
 
 /**
  * Public-facing abstraction that hides the internal Dagger component `BaseCriOrchestratorComponent`
